Updated old test results to contain whole number types instead of real number types where needed

+++ src/test/resources/results/testing/MyComponent10/testing_subpackage10_myComponent10_behaviorGeneration.h
@@ -12,7 +12,7 @@ double currentPositionX;
 double currentPositionY;
 double currentDirectionX;
 double currentDirectionY;
-double plannedTrajectoryLength;
+int plannedTrajectoryLength;
 Matrix plannedTrajectoryX;
 Matrix plannedTrajectoryY;
 double desiredDirectionX;
